requirements.txt
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
asgiref==3.11.0
boto3==1.42.30
botocore==1.42.30
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.1
cryptography==46.0.3
deprecation==2.1.0
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==40.1.2
fsspec==2026.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
jmespath==1.0.1
mmh3==5.2.0
multidict==6.7.0
openpyxl==3.1.5
packaging==25.0
pillow==12.1.0
pluggy==1.6.0
postgrest==2.27.1
pprintpp==0.4.0
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==2.27.1
reportlab==4.4.9
requests==2.32.5
rinoh-typeface-dejavuserif==0.1.3
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.5.5
roman-numerals==4.1.0
s3transfer==0.16.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.5
storage3==2.27.1
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.1
supabase-auth==2.27.1
supabase-functions==2.27.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
whitenoise==6.11.0
yarl==1.22.0
